backend/cloud_functions/sample_triage_agent/requirements.txt
google-cloud-bigquery>=3.12.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
functions-framework>=3.4.0
